Add DumbAlgo as new entry points with tests
Why: with the new way the Factory class handles its typenames, leaving
DumbAlgo inside conftest.py lead to a lot of errors that were not
warranted. By moving DumbAlgo to its own dependency like Gradient
Descent, we make sure it follows the `orion.algo` pattern
