+++ tests/dumbalgo/setup.cfg
@@ -1,0 +1,8 @@
+[aliases]
+test=pytest
+
+[bdist_wheel]
+universal=1
+
+[metadata]
+license_file=LICENSE
