+++ tox.ini
@@ -36,6 +36,7 @@ deps =
 usedevelop = True
 commands =
     pip install -U {toxinidir}/tests/functional/gradient_descent_algo
+    pip install -U {toxinidir}/tests/dumbalgo/
     coverage run --parallel-mode setup.py test --addopts '--timeout=180 {posargs}'
     coverage combine
     coverage report -m
@@ -81,6 +82,7 @@ deps =
 usedevelop = True
 commands =
     pip install -U {toxinidir}/tests/functional/gradient_descent_algo
+    pip install -U {toxinidir}/tests/dumbalgo
     python setup.py test --addopts '-vvv --exitfirst --capture=no --looponfail {posargs}'
 
 ## Configure linters
